[BE] Feat : 공유 게시판 스크랩 목록 가져오기 api

+++ backend/src/main/java/com/ssafy/backend/shareBoard/repository/ShareBoardCommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShareBoardCommentRepository extends JpaRepository<ShareBoardComment, Long> {
 
-    Long countByShareBoardSeqAndIsDeletedFalse(Long shareBoardSeq);
+    int countByShareBoardSeqAndIsDeletedFalse(Long shareBoardSeq);
 
 }
